+++ src/main/kotlin/com/github/tsdaer/dreamshaderlanguagesupport/language/lexer/DreamShaderLexer.kt
@@ -236,7 +236,7 @@ class DreamShaderLexer : LexerBase() {
             "bvec2", "bvec3", "bvec4",
             "mat2", "mat3", "mat4",
             "texture2d", "texturecube", "texture2darray", "texture3d", "volumetexture", "samplerstate",
-            "materialattributes",
+            "materialattributes", "substrate",
             "scalarparameter", "vectorparameter", "doublevectorparameter",
             "staticboolparameter", "staticswitchparameter",
             "textureobjectparameter",